Add prototype that uses in-browser live API calls
The labor action data is coming from a Grist document, rather than a Google spreadsheet.

This version requires an API proxy since we don't want to store Authorization headers in client-side code.

This version would not require a build step, but I have concerns about the safety and performance of the live proxy.