Fix inaccurate action card sort in the listing
This is done by adding the action._order property during the data file build rather than in the client.

The SQL statement sorts by start date, but once the actions are added to a Map, they are not ordered. So the _order property is calculated before any data is added to a Map, on the first iteration of the JSON data from the SQL results.

The _order property was probably in the client script because originally the SQL data was retrieved directly there.